requirements.txt
click==6.7
confluent-kafka==0.11.4
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.10
kafka-python==1.4.1
kazoo==2.4.0

numpy==1.14.1
opencv-python==3.4.0.12
six==1.11.0
tabulate==0.8.2
Werkzeug==0.14.1
wheel==0.26.0